♻️(chat) move agent to a dedicated module
This is a first step to tidy code and ease agent definition
readability.
